requirements.txt
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
googlemaps==4.6.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6